// src/server/model/itemGroupWiseTradingAccountReport.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ItemGroupWiseTradingAccountReportLine } from './itemGroupWiseTradingAccountReportLine';

export interface ItemGroupWiseTradingAccountReport { 
    reportLines?: Array<ItemGroupWiseTradingAccountReportLine>;
    dateFrom?: Date;
    dateTo?: Date;
    totalProfit?: number;
    totalPurchaseValue?: number;
    totalOpeningValue?: number;
    totalSaleValue?: number;
}